+++ client-modern/src/main/java/com/paraguacraft/pvp/modern/ParaguacraftPvPModern.java
@@ -17,7 +17,7 @@ import net.minecraft.client.MinecraftClient;
 public class ParaguacraftPvPModern implements ClientModInitializer {
 
     public static final String MOD_ID = "paraguacraftpvp-modern";
-    public static final String VERSION = "0.6.3";
+    public static final String VERSION = "0.6.4";
 
     @Override
     public void onInitializeClient() {